// src/api/interfaces/IEmployee.interface.ts
export interface IEmployee {
  employee_id: string;
  employee_surname: string;
  employee_name: string;
  employee_patronymic: string;
  employee_role: number;
  employee_salary: number;
  employee_start_date: string;
  employee_birth_date: string;
  employee_phone_number: string;
  employee_city: string;
  employee_street: string;
  employee_zip_code: string;
}